Extract class 'Emote' to separate file
